Add AttestationSlashingBuilder
