+++ eth2/types/src/attester_slashing.rs
@@ -1,8 +1,12 @@
-use crate::{test_utils::TestRandom, ChainSpec, SlashableAttestation};
+use crate::{test_utils::TestRandom, SlashableAttestation};
 use rand::RngCore;
 use serde_derive::Serialize;
 use ssz_derive::{Decode, Encode, TreeHash};
 use test_random_derive::TestRandom;
+
+mod builder;
+
+pub use builder::AttesterSlashingBuilder;
 
 #[derive(Debug, PartialEq, Clone, Serialize, Encode, Decode, TreeHash, TestRandom)]
 pub struct AttesterSlashing {
